+++ aiobungie/client.py
@@ -31,13 +31,23 @@ __all__: list[str] = ["Client"]
 import asyncio
 import typing
 
-from aiobungie import crate
 from aiobungie import rest as rest_
-from aiobungie.ext import Manifest
+from aiobungie.crate import user
+from aiobungie.ext import meta
 from aiobungie.internal import enums
 from aiobungie.internal import factory
 from aiobungie.internal import helpers
 from aiobungie.internal import traits
+
+if typing.TYPE_CHECKING:
+    from aiobungie.crate import activity
+    from aiobungie.crate import application
+    from aiobungie.crate import character
+    from aiobungie.crate import clans
+    from aiobungie.crate import entity
+    from aiobungie.crate import friends
+    from aiobungie.crate import milestones
+    from aiobungie.crate import profile
 
 
 class Client(traits.ClientBase):
@@ -93,7 +103,7 @@ class Client(traits.ClientBase):
 
     # * Unspecified methods. *#
 
-    async def fetch_manifest(self) -> Manifest:
+    async def fetch_manifest(self) -> meta.Manifest:
         """Access The bungie Manifest.
 
         Returns
@@ -101,11 +111,11 @@ class Client(traits.ClientBase):
         `aiobungie.ext.Manifest`
             A Manifest crate.
         """
-        return Manifest(self._token)
+        return meta.Manifest(self._token)
 
     # * User methods.
 
-    async def fetch_own_bungie_user(self, *, access_token: str) -> crate.user.User:
+    async def fetch_own_bungie_user(self, *, access_token: str) -> user.User:
         """Fetch and return a user object of the bungie net user associated with account.
 
         .. warning::
@@ -126,7 +136,7 @@ class Client(traits.ClientBase):
         assert isinstance(resp, dict)
         return self.serialize.deserialize_user(resp)
 
-    async def fetch_user(self, id: int) -> crate.user.BungieUser:
+    async def fetch_user(self, id: int) -> user.BungieUser:
         """Fetch a Bungie user by their id.
 
         .. note::
@@ -152,7 +162,7 @@ class Client(traits.ClientBase):
         assert isinstance(payload, dict)
         return self.serialize.deserialize_bungie_user(payload)
 
-    async def search_users(self, name: str, /) -> typing.Sequence[crate.DestinyUser]:
+    async def search_users(self, name: str, /) -> typing.Sequence[user.DestinyUser]:
         """Search for players and return all players that matches the same name.
 
         Parameters
@@ -169,7 +179,7 @@ class Client(traits.ClientBase):
         assert isinstance(payload, dict)
         return self.serialize.deseialize_found_users(payload)
 
-    async def fetch_user_themes(self) -> typing.Sequence[crate.user.UserThemes]:
+    async def fetch_user_themes(self) -> typing.Sequence[user.UserThemes]:
         """Fetch all available user themes.
 
         Returns
@@ -186,7 +196,7 @@ class Client(traits.ClientBase):
         credential: int,
         type: helpers.IntAnd[enums.CredentialType] = enums.CredentialType.STEAMID,
         /,
-    ) -> crate.user.HardLinkedMembership:
+    ) -> user.HardLinkedMembership:
         """Gets any hard linked membership given a credential.
         Only works for credentials that are public just `aiobungie.CredentialType.STEAMID` right now.
         Cross Save aware.
@@ -210,7 +220,7 @@ class Client(traits.ClientBase):
         payload = await self.rest.fetch_hard_linked(credential, type)
         assert isinstance(payload, dict)
 
-        return crate.user.HardLinkedMembership(
+        return user.HardLinkedMembership(
             id=int(payload["membershipId"]),
             type=enums.MembershipType(payload["membershipType"]),
             cross_save_type=enums.MembershipType(payload["CrossSaveOverriddenType"]),
@@ -221,7 +231,7 @@ class Client(traits.ClientBase):
         id: int,
         type: helpers.IntAnd[enums.MembershipType] = enums.MembershipType.NONE,
         /,
-    ) -> crate.User:
+    ) -> user.User:
         """Fetch Bungie user's memberships from their id.
 
         Notes
@@ -259,7 +269,7 @@ class Client(traits.ClientBase):
         memberid: int,
         type: helpers.IntAnd[enums.MembershipType],
         /,
-    ) -> crate.Profile:
+    ) -> profile.Profile:
         """
         Fetche a bungie profile.
 
@@ -293,7 +303,7 @@ class Client(traits.ClientBase):
         /,
         *,
         all: bool = False,
-    ) -> crate.profile.LinkedProfile:
+    ) -> profile.LinkedProfile:
         """Returns a summary information about all profiles linked to the requested member.
 
         The passed membership id/type maybe a Bungie.Net membership or a Destiny memberships.
@@ -331,7 +341,7 @@ class Client(traits.ClientBase):
         name: str,
         type: helpers.IntAnd[enums.MembershipType] = enums.MembershipType.ALL,
         /,
-    ) -> typing.Sequence[crate.user.DestinyUser]:
+    ) -> typing.Sequence[user.DestinyUser]:
         """Fetch a Destiny 2 Player.
 
         .. note::
@@ -367,7 +377,7 @@ class Client(traits.ClientBase):
         memberid: int,
         type: helpers.IntAnd[enums.MembershipType],
         character: enums.Class,
-    ) -> crate.Character:
+    ) -> character.Character:
         """Fetch a Destiny 2 character.
 
         Parameters
@@ -407,7 +417,7 @@ class Client(traits.ClientBase):
         *,
         page: int = 1,
         limit: int = 1,
-    ) -> crate.activity.Activity:
+    ) -> activity.Activity:
         """Fetch a Destiny 2 activity for the specified user id and character.
 
         Parameters
@@ -449,9 +459,7 @@ class Client(traits.ClientBase):
         assert isinstance(resp, dict)
         return self.serialize.deserialize_activity(resp)
 
-    async def fetch_post_activity(
-        self, instance: int, /
-    ) -> crate.activity.PostActivity:
+    async def fetch_post_activity(self, instance: int, /) -> activity.PostActivity:
         """Fetch a post activity details.
 
         .. warning::
@@ -474,7 +482,7 @@ class Client(traits.ClientBase):
 
     # * Destiny 2 Clans or GroupsV2.
 
-    async def fetch_clan_from_id(self, id: int, /) -> crate.Clan:
+    async def fetch_clan_from_id(self, id: int, /) -> clans.Clan:
         """Fetch a Bungie Clan by its id.
 
         Parameters
@@ -498,7 +506,7 @@ class Client(traits.ClientBase):
 
     async def fetch_clan(
         self, name: str, /, type: helpers.IntAnd[enums.GroupType] = enums.GroupType.CLAN
-    ) -> crate.Clan:
+    ) -> clans.Clan:
         """Fetch a Clan by its name.
         This method will return the first clan found with given name.
 
@@ -525,14 +533,14 @@ class Client(traits.ClientBase):
 
     async def fetch_clan_conversations(
         self, clan_id: int, /
-    ) -> typing.Sequence[crate.clans.ClanConversation]:
+    ) -> typing.Sequence[clans.ClanConversation]:
         resp = await self.rest.fetch_clan_conversations(clan_id)
         assert isinstance(resp, list)
         return self.serialize.deserialize_clan_convos(resp)
 
     async def fetch_clan_admins(
         self, clan_id: int, /
-    ) -> typing.Sequence[crate.clans.ClanAdmin]:
+    ) -> typing.Sequence[clans.ClanAdmin]:
         """Fetch the clan founder and admins.
 
         Parameters
@@ -562,7 +570,7 @@ class Client(traits.ClientBase):
         *,
         filter: int = 0,
         group_type: enums.GroupType = enums.GroupType.CLAN,
-    ) -> typing.Optional[crate.clans.GroupMember]:
+    ) -> typing.Optional[clans.GroupMember]:
         """Fetch information about the groups that a given member has joined.
 
         Parameters
@@ -599,7 +607,7 @@ class Client(traits.ClientBase):
         *,
         filter: int = 0,
         group_type: helpers.IntAnd[enums.GroupType] = enums.GroupType.CLAN,
-    ) -> typing.Optional[crate.clans.GroupMember]:
+    ) -> typing.Optional[clans.GroupMember]:
         resp = await self.rest.fetch_potential_groups_for_member(
             member_id, member_type, filter=filter, group_type=group_type
         )
@@ -612,7 +620,7 @@ class Client(traits.ClientBase):
         name: typing.Optional[str] = None,
         type: helpers.IntAnd[enums.MembershipType] = enums.MembershipType.NONE,
         /,
-    ) -> crate.clans.ClanMember:
+    ) -> clans.ClanMember:
         """Fetch a Bungie Clan member.
 
         .. note::
@@ -654,7 +662,7 @@ class Client(traits.ClientBase):
         clan_id: int,
         type: helpers.IntAnd[enums.MembershipType] = enums.MembershipType.NONE,
         /,
-    ) -> typing.Sequence[crate.clans.ClanMember]:
+    ) -> typing.Sequence[clans.ClanMember]:
         """Fetch a Bungie Clan member. if no members found in the clan
         you will get an empty sequence.
 
@@ -686,7 +694,7 @@ class Client(traits.ClientBase):
         assert isinstance(resp, dict)
         return self.serialize.deserialize_clan_members(resp)
 
-    async def fetch_clan_banners(self) -> typing.Sequence[crate.clans.ClanBanner]:
+    async def fetch_clan_banners(self) -> typing.Sequence[clans.ClanBanner]:
         """Fetch the clan banners.
 
         Returns
@@ -706,7 +714,7 @@ class Client(traits.ClientBase):
         group_id: int,
         membership_id: int,
         membership_type: helpers.IntAnd[enums.MembershipType],
-    ) -> crate.Clan:
+    ) -> clans.Clan:
         """Kick a member from the clan.
 
         .. note::
@@ -739,7 +747,7 @@ class Client(traits.ClientBase):
 
     # * Destiny 2 Definitions. Entities.
 
-    async def fetch_inventory_item(self, hash: int, /) -> crate.entity.InventoryEntity:
+    async def fetch_inventory_item(self, hash: int, /) -> entity.InventoryEntity:
         """Fetch a static inventory item entity given a its hash.
 
         Parameters
@@ -759,7 +767,53 @@ class Client(traits.ClientBase):
     # * These methods should be for Special bungie endpoints, i.e,
     # * Applications, Forums, Polls, Trending, etc.
 
-    async def fetch_app(self, appid: int, /) -> crate.Application:
+    async def fetch_friends(
+        self, access_token: str, /
+    ) -> typing.Sequence[friends.Friend]:
+        """Fetch bungie friend list.
+
+        .. note::
+            This requests OAuth2: ReadUserData scope.
+
+        Parameters
+        -----------
+        access_token : `str`
+            The bearer access token associated with the bungie account.
+
+        Returns
+        -------
+        `typing.Sequence[aiobungie.crate.Friend]`
+            A sequence of the found friends.
+        """
+
+        resp = await self.rest.fetch_friends(access_token)
+        assert isinstance(resp, dict)
+        return self.serialize.deserialize_friends(resp)
+
+    async def fetch_friend_requests(
+        self, access_token: str, /
+    ) -> friends.FriendRequestView:
+        """Fetch pending bungie friend requests queue.
+
+        .. note::
+            This requests OAuth2: ReadUserData scope.
+
+        Parameters
+        -----------
+        access_token : `str`
+            The bearer access token associated with the bungie account.
+
+        Returns
+        -------
+        `aiobungie.crate.FriendRequestView`
+            A friend requests view object includes a sequence of incoming and outgoing requests.
+        """
+
+        resp = await self.rest.fetch_friend_requests(access_token)
+        assert isinstance(resp, dict)
+        return self.serialize.deserialize_friend_requests(resp)
+
+    async def fetch_app(self, appid: int, /) -> application.Application:
         """Fetch a Bungie Application.
 
         Parameters
@@ -778,7 +832,7 @@ class Client(traits.ClientBase):
 
     async def fetch_public_milestone_content(
         self, milestone_hash: int, /
-    ) -> crate.milestones.Milestone:
+    ) -> milestones.Milestone:
         """Fetch the milestone content given its hash.
 
         Parameters